scripts/requirements.txt
pycryptodome>=3.19.0
pywin32>=306
httpx>=0.25.0
